Fix root path issues
