+++ .trivyignore
@@ -31,6 +31,7 @@ CVE-2022-37616
 CVE-2022-38900
 CVE-2022-39353
 CVE-2023-0286
+CVE-2023-28155
 DS001
 DS002
 DS026
