+++ docs/schema/sfdx-hardis-json-schema-parameters.html
@@ -4202,6 +4202,6 @@
 </div>
 
     <footer>
-        <p class="generated-by-footer">Generated using <a href="https://github.com/coveooss/json-schema-for-humans">json-schema-for-humans</a> on 2023-03-16 at 11:48:45 +0100</p>
+        <p class="generated-by-footer">Generated using <a href="https://github.com/coveooss/json-schema-for-humans">json-schema-for-humans</a> on 2023-03-20 at 11:16:36 +0100</p>
     </footer></body>
 </html>
