+++ package.json
@@ -22,6 +22,7 @@
     "@typescript-eslint/eslint-plugin": "^4.17.0",
     "@typescript-eslint/parser": "^4.17.0",
     "@xmldom/xmldom": "^0.8.6",
+    "app-root-path": "^3.1.0",
     "axios": "^0.21.1",
     "chalk": "^4.1.0",
     "changed-git-files": "^0.0.1",
