+++ src/common/gitProvider/index.ts
@@ -4,7 +4,6 @@ import { AzureDevopsProvider } from "./azureDevops";
 import { GithubProvider } from "./github";
 import { GitlabProvider } from "./gitlab";
 import { GitProviderRoot } from "./gitProviderRoot";
-import { PullRequestMessageRequest } from "./types/gitProvider";
 
 export abstract class GitProvider {
   static getInstance(): GitProviderRoot {
@@ -63,3 +62,15 @@ export abstract class GitProvider {
     }
   }
 }
+
+export declare type PullRequestMessageRequest = {
+  title: string;
+  message: string;
+  messageKey: string;
+  status: ["valid", "invalid", "tovalidate"];
+};
+
+export declare type PullRequestMessageResult = {
+  posted: boolean;
+  providerResult: any;
+};
