--- src/common/gitProvider/types/gitProvider.d.ts
@@ -1,11 +1,0 @@
-export declare type PullRequestMessageRequest = {
-  title: string;
-  message: string;
-  messageKey: string;
-  status: ["valid", "invalid", "tovalidate"];
-};
-
-export declare type PullRequestMessageResult = {
-  posted: boolean;
-  providerResult: any;
-};
