+++ src/common/keyValueProviders/salesforce.ts
@@ -1,4 +1,5 @@
 import { Connection } from "@salesforce/core";
+import * as appRootPath from 'app-root-path';
 import * as c from "chalk";
 import * as he from "he";
 import * as path from "path";
@@ -86,7 +87,7 @@ export class SalesforceProvider implements KeyValueProviderInterface {
     // Deploy KeyValueStore object on DevHub org
     try {
       await deployMetadatas({
-        deployDir: path.join(path.join(__dirname, "../../../defaults/utils/sfdxHardisKeyValueStore", ".")),
+        deployDir: path.join(path.join(appRootPath.toString(), "defaults/utils/sfdxHardisKeyValueStore", ".")),
         soap: true,
         targetUsername: options.devHubConn.options.authInfo.fields.username,
       });
